+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 * **utils:** Add datetime_format utility for standardizing datetime formatting
+* **ses:** Improve SES identity module with datetime formatting and add tests
 
 ## [1.1.0](https://github.com/KKamJi98/listup_aws_resources/compare/v1.0.0...v1.1.0) (2025-07-02)